// constants/types.ts
  export type Canteens = {
    alias?: string | null;
    building?: string | Buildings | null;
    date_created?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    foodoffers_import_delete_all_without_dates?: boolean | null;
    foodoffers_import_without_date?: boolean | null;
    foodservice_hours: any[] | CanteensFoodserviceHours[];
    foodservice_hours_during_semester_break: any[] | CanteensFoodserviceHoursDuringSemesterBreak[];
    id: string;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
    utilization_group?: string | UtilizationsGroups | null;
  };
// ...
  export type DirectusAccess = {
    id: string;
    policy: string | DirectusPolicies;
    role?: string | DirectusRoles | null;
    sort?: number | null;
    user?: string | DirectusUsers | null;
  };
// ...
  export type DirectusFiles = {
    charset?: string | null;
    created_on: string;
    description?: string | null;
    duration?: number | null;
    embed?: string | null;
    filename_disk?: string | null;
    filename_download: string;
    filesize?: number | null;
    focal_point_x?: number | null;
    focal_point_y?: number | null;
    folder?: string | DirectusFolders | null;
    height?: number | null;
    id: string;
    location?: string | null;
    metadata?: unknown | null;
    modified_by?: string | DirectusUsers | null;
    modified_on: string;
    storage: string;
    tags?: unknown | null;
    title?: string | null;
    tus_data?: unknown | null;
    tus_id?: string | null;
    type?: string | null;
    uploaded_by?: string | DirectusUsers | null;
    uploaded_on?: string | null;
    width?: number | null;
  };
// ...
  export type DirectusFolders = {
    id: string;
    name: string;
    parent?: string | DirectusFolders | null;
  };
// ...
  export type DirectusRoles = {
    children: any[] | DirectusRoles[];
    description?: string | null;
    icon: string;
    id: string;
    name: string;
    parent?: string | DirectusRoles | null;
    policies: any[] | DirectusAccess[];
    users: any[] | DirectusUsers[];
    users_group: string;
  };
// ...
  export type DirectusUsers = {
    appearance?: string | null;
    auth_data?: unknown | null;
    avatar?: string | DirectusFiles | null;
    description?: string | null;
    email?: string | null;
    email_notifications?: boolean | null;
    external_identifier?: string | null;
    first_name?: string | null;
    id: string;
    language?: string | null;
    last_access?: string | null;
    last_name?: string | null;
    last_page?: string | null;
    location?: string | null;
    password?: string | null;
    policies: any[] | DirectusAccess[];
    profile?: string | Profiles | null;
    provider: string;
    role?: string | DirectusRoles | null;
    status: string;
    tags?: unknown | null;
    tfa_secret?: string | null;
    theme_dark?: string | null;
    theme_dark_overrides?: unknown | null;
    theme_light?: string | null;
    theme_light_overrides?: unknown | null;
    title?: string | null;
    token?: string | null;
  };
// ...
  export type Foodoffers = {
    alias?: string | null;
    calories_kcal?: number | null;
    canteen?: string | Canteens | null;
    carbohydrate_g?: number | null;
    category?: string | null;
    co2_g?: number | null;
    co2_rating?: string | null;
    co2_saving_percentage?: number | null;
    date?: string | null;
    date_created?: string | null;
    date_updated?: string | null;
    environmental_impact: string;
    fat_g?: number | null;
    fiber_g?: number | null;
    food?: string | Foods | null;
    foodoffer_category?: string | FoodoffersCategories | null;
    id: string;
    markings: any[] | FoodoffersMarkings[];
    nutrition: string;
    price_employee?: number | null;
    price_guest?: number | null;
    price_student?: number | null;
    prices: string;
    protein_g?: number | null;
    salt_g?: number | null;
    saturated_fat_g?: number | null;
    sort?: number | null;
    status?: string | null;
    sugar_g?: number | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
// ...
  export type Foods = {
    alias?: string | null;
    calories_kcal?: number | null;
    carbohydrate_g?: number | null;
    category?: string | null;
    co2_g?: number | null;
    co2_rating?: string | null;
    co2_saving_percentage?: number | null;
    date_created?: string | null;
    date_updated?: string | null;
    environmental_impact: string;
    extra?: string | null;
    fat_g?: number | null;
    feedbacks: any[] | FoodsFeedbacks[];
    fiber_g?: number | null;
    food_category?: string | FoodsCategories | null;
    id: string;
    image?: string | DirectusFiles | null;
    image_remote_url?: string | null;
    image_thumb_hash?: string | null;
    markings: any[] | FoodsMarkings[];
    nutrition: string;
    protein_g?: number | null;
    rating_amount?: number | null;
    rating_amount_legacy?: number | null;
    rating_average?: number | null;
    rating_average_legacy?: number | null;
    rating_legacy_settings: string;
    rating_settings: string;
    salt_g?: number | null;
    saturated_fat_g?: number | null;
    sort?: number | null;
    status?: string | null;
    sugar_g?: number | null;
    translations: any[] | FoodsTranslations[];
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
  
  export type FoodsCategories = {
    alias?: string | null;
    date_created?: string | null;
    date_updated?: string | null;
    external_identifier?: string | null;
    id: string;
    sort?: number | null;
    status: string;
    translations: any[] | FoodsCategoriesTranslations[];
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
  
  export type FoodsCategoriesTranslations = {
    be_source_for_translations?: boolean | null;
    foods_categories_id?: string | FoodsCategories | null;
    id: number;
    languages_code?: string | Languages | null;
    let_be_translated?: boolean | null;
    name?: string | null;
    translation_settings: string;
  };
  
  export type FoodsFeedbacks = {
    canteen?: string | Canteens | null;
    comment?: string | null;
    date_created?: string | null;
    date_updated?: string | null;
    food?: string | Foods | null;
    foodoffer?: string | Foodoffers | null;
    id: string;
    notify?: boolean | null;
    profile?: string | Profiles | null;
    rating?: number | null;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
// ...
  export type FoodsMarkings = {
    foods_id?: string | Foods | null;
    id: number;
    markings_id?: string | Markings | null;
  };
  
  export type FoodsTranslations = {
    be_source_for_translations?: boolean | null;
    foods_id?: string | Foods | null;
    id: number;
    languages_code?: string | Languages | null;
    let_be_translated?: boolean | null;
    name?: string | null;
    translation_settings: string;
  };
  
  export type Languages = {
    code: string;
    direction?: string | null;
    name?: string | null;
    status?: string | null;
  };
// ...
  export type Markings = {
    alias?: string | null;
    background_color?: string | null;
    date_created?: string | null;
    date_updated?: string | null;
    excluded_by_markings: any[] | MarkingsExclusions[];
    external_identifier?: string | null;
    group?: string | MarkingsGroups | null;
    hide_border?: boolean | null;
    icon?: string | null;
    id: string;
    image?: string | DirectusFiles | null;
    image_remote_url?: string | null;
    image_thumb_hash?: string | null;
    invert_background_color?: boolean | null;
    short_code?: string | null;
    show_on_card?: boolean | null;
    sort?: number | null;
    status?: string | null;
    translations: any[] | MarkingsTranslations[];
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
// ...
  export type Profiles = {
    avatar?: unknown | null;
    buildings_favorites: any[] | ProfilesBuildingsFavorites[];
    buildings_last_opened: any[] | ProfilesBuildingsLastOpened[];
    canteen?: string | Canteens | null;
    course_timetable?: string | null;
    credit_balance?: number | null;
    credit_balance_date_updated?: string | null;
    credit_balance_last_transaction?: number | null;
    date_created?: string | null;
    date_privacy_policy_accepted?: string | null;
    date_updated?: string | null;
    devices: any[] | Devices[];
    foods_feedbacks: any[] | FoodsFeedbacks[];
    id: string;
    language?: string | Languages | null;
    markings: any[] | ProfilesMarkings[];
    nickname?: string | null;
    notifiy_on_free_apartments?: boolean | null;
    price_group?: string | null;
    sort?: number | null;
    status?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };